ai-api/requirements.txt
flask
flask-cors
tensorflow
tf-keras
numpy
pillow
opencv-python
matplotlib
ultralytics
deepface
scikit-learn
